added the old version of Max's code, I haven't received the new stuff yet
