--- VR_Piano/Assets/NoteCallback.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ef80b44c05b194c4c8d2a3405558a3c2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
