+++ VR_Piano/Assets/Scenes/SampleScene.unity
@@ -713,7 +713,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   KeyPreFab: {fileID: 3101965913632677811, guid: bd03f546458884001b8498565c2bee96, type: 3}
-  KeyCount: 73
+  KeyCount: 120
   spacing: 0.01
 --- !u!1001 &453753817
 PrefabInstance:
@@ -4184,3 +4184,4 @@ SceneRoots:
   - {fileID: 3001774293689589082}
   - {fileID: 2073726594}
   - {fileID: 296496623}
+  - {fileID: 722842072}
