+++ VR_Piano/Assets/Scripts/NoteCallback.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 804701c3b282e7e46819d68174dcbe4c
+guid: ef80b44c05b194c4c8d2a3405558a3c2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
